+++ src/faceAgent.h
@@ -15,7 +15,7 @@ struct PersonFace;
 class UserFace {
  public:
   UserFace(const std::string &userId, 
-           const std::string &userName);
+           	   const std::string &userName);
   int addImageFace(std::shared_ptr <ImageFace> imageFace);
   int delImageFace(const std::string &faceToken);
   std::shared_ptr<ImageFace>  getImageFace(const std::string &faceToken);
@@ -35,8 +35,7 @@ class UserFace {
 class GroupFace {
  public:
   GroupFace(const std::string &groupId);
-  int addUserFace(const std::string &userId, const std::string &userName);
-  int delUserFace(const std::string &userId);
+  int addUser(const std::string &userId, const std::string &userName);
   int delUser(const std::string &userId);
   std::shared_ptr<UserFace> getUserFace(const std::string &userId);
   std::map<std::string, std::shared_ptr<UserFace>> getUserFaces() {
@@ -83,8 +82,6 @@ class FaceAgent {
   private:
     std::map<std::string, std::shared_ptr<AppFace>> appFaces;
 };
-
-
 
 }
 #endif
